requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
alembic==1.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
aiofiles==0.7.0
python-dotenv==0.19.0
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0